requirements.txt
notebook
PyYAML
pillow
opencv-python
imutils
-e .